add attributes package
